rename TGeo to Root plugin
