+++ Examples/Python/CMakeLists.txt
@@ -110,8 +110,8 @@ else()
     target_sources(ActsPythonBindings PRIVATE src/GeoModelStub.cpp)
 endif()
 
-if(ACTS_BUILD_PLUGIN_TGEO)
-    target_link_libraries(ActsPythonBindings PUBLIC ActsPluginTGeo)
+if(ACTS_BUILD_PLUGIN_ROOT)
+    target_link_libraries(ActsPythonBindings PUBLIC ActsPluginRoot)
     target_sources(ActsPythonBindings PRIVATE src/TGeo.cpp)
 else()
     target_sources(ActsPythonBindings PRIVATE src/TGeoStub.cpp)
