+++ Plugins/Root/include/Acts/Plugins/Root/ITGeoIdentifierProvider.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "Acts/Geometry/GeometryContext.hpp"
-#include "Acts/Plugins/TGeo/TGeoDetectorElement.hpp"
+#include "Acts/Plugins/Root/TGeoDetectorElement.hpp"
 
 class TGeoNode;
