+++ Plugins/Root/include/Acts/Plugins/Root/TGeoLayerBuilder.hpp
@@ -15,7 +15,7 @@
 #include "Acts/Geometry/LayerCreator.hpp"
 #include "Acts/Geometry/ProtoLayerHelper.hpp"
 #include "Acts/Geometry/SurfaceBinningMatcher.hpp"
-#include "Acts/Plugins/TGeo/ITGeoIdentifierProvider.hpp"
+#include "Acts/Plugins/Root/ITGeoIdentifierProvider.hpp"
 #include "Acts/Utilities/BinningType.hpp"
 #include "Acts/Utilities/Logger.hpp"
