+++ Plugins/Root/src/TGeoMaterialConverter.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/TGeo/TGeoMaterialConverter.hpp"
+#include "Acts/Plugins/Root/TGeoMaterialConverter.hpp"
 
 #include "Acts/Material/Material.hpp"
