+++ Plugins/Root/src/TGeoPrimitivesHelper.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/TGeo/TGeoPrimitivesHelper.hpp"
+#include "Acts/Plugins/Root/TGeoPrimitivesHelper.hpp"
 
 bool Acts::TGeoPrimitivesHelper::match(const char* first, const char* second) {
   // If we reach at the end of both strings, we are done
