+++ Tests/UnitTests/Plugins/CMakeLists.txt
@@ -5,7 +5,7 @@ add_subdirectory_if(ExaTrkX ACTS_BUILD_PLUGIN_EXATRKX)
 add_subdirectory_if(Geant4 ACTS_BUILD_PLUGIN_GEANT4)
 add_subdirectory_if(GeoModel ACTS_BUILD_PLUGIN_GEOMODEL)
 add_subdirectory_if(Json ACTS_BUILD_PLUGIN_JSON)
-add_subdirectory_if(TGeo ACTS_BUILD_PLUGIN_TGEO)
+add_subdirectory_if(TGeo ACTS_BUILD_PLUGIN_ROOT)
 add_subdirectory_if(EDM4hep ACTS_BUILD_PLUGIN_EDM4HEP)
 add_subdirectory_if(FpeMonitoring ACTS_BUILD_PLUGIN_FPEMON)
 add_subdirectory_if(Podio ACTS_BUILD_PLUGIN_PODIO)
