+++ Tests/UnitTests/Plugins/TGeo/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(unittest_extra_libraries ActsPluginTGeo)
+set(unittest_extra_libraries ActsPluginRoot)
 
 add_unittest(TGeoArb8Conversion TGeoArb8ConversionTests.cpp)
 add_unittest(TGeoBBoxConversion TGeoBBoxConversionTests.cpp)
