+++ Tests/UnitTests/Plugins/TGeo/TGeoPrimitivesHelperTests.cpp
@@ -8,7 +8,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "Acts/Plugins/TGeo/TGeoPrimitivesHelper.hpp"
+#include "Acts/Plugins/Root/TGeoPrimitivesHelper.hpp"
 
 #include <string>
 #include <vector>
