+++ cmake/ActsConfig.cmake.in
@@ -62,7 +62,7 @@ endif()
 if(PluginJson IN_LIST Acts_COMPONENTS)
   find_dependency(nlohmann_json @nlohmann_json_VERSION@ CONFIG EXACT)
 endif()
-if(PluginTGeo IN_LIST Acts_COMPONENTS)
+if(PluginRoot IN_LIST Acts_COMPONENTS)
   find_dependency(ROOT @ROOT_VERSION@ CONFIG EXACT)
 endif()
 if(PluginActSVG IN_LIST Acts_COMPONENTS)
