+++ docs/plugins/dd4hep.md
@@ -98,7 +98,7 @@ changed in ACTS.
 The DD4hepPlugin is only build on demand. The DD4hepPlugin depends on the
 TGeoPlugin therefore both plugins need to be installed. During the cmake
 configuration the flags `ACTS_BUILD_PLUGIN_DD4HEP=on` and
-`ACTS_BUILD_PLUGIN_TGEO=on` need to be set. In addition, ROOT and DD4hep
+`ACTS_BUILD_PLUGIN_ROOT=on` need to be set. In addition, ROOT and DD4hep
 installations need to be available to cmake.
 
 ## Prerequisites
